Implement import of account 3224 and 3241 (issue #9)
This makes it possible to import transfers from other WOs and other
organizations. However, there are several limitations: often the amount
of entries in these categories is after deducting the assessment, so it
is not clear what the original donation amount was. It is possible that
one entry represents multiple donations. And often the donations are in
foreign currency so there's the problem of conversion rates. The
statement doesn't state what the original amount in foreign currency was,
so we can only work with the converted EUR amount.