Improve handling of donor names split accross two lines
If the donor last name, first name and spouse name is too long to
fit in one line the statement splits them on two lines. This change
handles the cases where the spouse name is on the second line with
the separator being the last word on the first line or the first
word on the second line.